🛡️ Sentinel: [HIGH] Fix XSS vulnerability in DomSanitizer bypassing
Created a centralized utility function (`isTrustedUrl`) to consistently validate URLs against predefined trusted hosts. Applied this to `MediaPreviewComponent` and `MessageMediaContentComponent` to prevent XSS caused by bypassing security on unvalidated `mediaData.link` inputs, falling back safely when untrusted URLs are passed.
