+++ src/app/contacts/contact-info/media-preview/media-preview.component.ts
@@ -11,6 +11,7 @@ import { UseMedia } from "../../../../core/message/model/media-data.model";
 import { MediaStoreService } from "../../../../core/media/store/media-store.service";
 import { MatIconModule } from "@angular/material/icon";
 import { NGXLogger } from "ngx-logger";
+import { isTrustedUrl } from "../../../../core/common/util/url-validator.util";
 
 @Component({
     selector: "app-media-preview",
@@ -31,7 +32,7 @@ export class MediaPreviewComponent implements OnInit {
 
     @Input() message!: Conversation;
 
-    mediaSafeUrl: SafeUrl = ""; // Safe URL for media
+    mediaSafeUrl: SafeUrl | string = ""; // Safe URL for media
     options = false;
 
     async ngOnInit(): Promise<void> {
@@ -43,7 +44,12 @@ export class MediaPreviewComponent implements OnInit {
         if (!mediaData) return;
         const url = mediaData?.link;
         if (url) {
-            this.mediaSafeUrl = this.sanitizer.bypassSecurityTrustUrl(url); // Sanitize the URL
+            if (isTrustedUrl(url)) {
+                this.mediaSafeUrl = this.sanitizer.bypassSecurityTrustUrl(url); // Sanitize the URL
+            } else {
+                this.logger.warn(`MediaPreviewComponent: Blocked untrusted URL: ${url}`);
+                this.mediaSafeUrl = "";
+            }
             return;
         }
         if (!mediaData.id) return;
